archive: write the archive to local system
When a `makefile` uses a `.archive` for the target of a rule, if we
don't write the file from `cas artifact` a subsequent `make` invocation
will re-run as the `.archive` file doesn't exist.

If however, you had deleted the output directory entirely, so that cas
downloads the artifacts, it would work!